// src/api/generated/models/patchCompaniesCompanyIdInvoicesInvoiceIdBody.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { PatchCompaniesCompanyIdInvoicesInvoiceIdBodyNotes } from './patchCompaniesCompanyIdInvoicesInvoiceIdBodyNotes';

export type PatchCompaniesCompanyIdInvoicesInvoiceIdBody = {
  /** @minLength 1 */
  number?: string;
  notes?: PatchCompaniesCompanyIdInvoicesInvoiceIdBodyNotes;
};